pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gobp"
version = "0.1.0"
description = "Graph of Brainstorm Project — long-term memory for AI agents"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "GoBP Authors" }
]
keywords = ["mcp", "knowledge-graph", "ai-memory", "ai-agents"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]

dependencies = [
    "mcp>=0.9.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]
# PostgreSQL v3 (GOBP_DB_URL) — optional so file-only installs stay dependency-light.
postgres = [
    "psycopg2-binary>=2.9.9",
]

[tool.pytest.ini_options]
addopts = ["-m", "not slow"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (batch 100+ nodes, > 5s, or perf benchmarks)",
    "postgres_v3: tests that require PostgreSQL with schema v3 (set GOBP_DB_URL)",
    "file_only: tests intended for file-backed graph only (no live DB assumptions)",
    "destructive: tests that TRUNCATE/DROP or otherwise wipe remote DB tables (extra guardrails)",
]

[project.scripts]
gobp = "gobp.cli.__main__:main"

[tool.setuptools.packages.find]
include = ["gobp*"]
exclude = ["tests*", "examples*"]

[tool.setuptools.package-data]
gobp = ["schema/*.yaml", "templates/*.md"]
